+++ .gitignore
@@ -134,4 +134,4 @@ model/saved/**
 data/saved/**
 data/Planetoid/
 
-data/article/
+data/movie/
